# improve chunk throughput with elide array bound checks