@MichaelEischer 2023-09-28
The change to use
```
		out := win[wpos%windowSize]
		win[wpos%windowSize] = b
```

yields the following speedup for me:

```
goos: linux
goarch: amd64
pkg: github.com/restic/chunker
cpu: AMD Ryzen 5 5600X 6-Core Processor
name                  old time/op   new time/op   delta
ChunkerWithSHA256-12   56.0ms ± 0%   54.5ms ± 0%  -2.60%  (p=0.000 n=8+9)
Chunker-12             41.5ms ± 0%   40.0ms ± 0%  -3.58%  (p=0.000 n=9+9)

name                  old speed     new speed     delta
ChunkerWithSHA256-12  599MB/s ± 0%  615MB/s ± 0%  +2.67%  (p=0.000 n=8+9)
Chunker-12            809MB/s ± 0%  840MB/s ± 0%  +3.71%  (p=0.000 n=9+9)
```

The other changes have no measurable effect. Regarding readability I'd prefer to keep the `slide()` method. Thus, please limit the changes in the PR to the just mentioned code snippet and update the corresponding comment to read `// limit wpos to elide array bound checks`